Consider one-off notifications to be jobs
When we get back the notifications for a service there is the option to
filter out those created by a job.

Previously this excluded one-off notifications, because they created a
job. Now that they are created directly by the API they are not getting
filtered out.

This commit changes the filter to consider a one-off notification to
still be created by a job because:
- it keeps the behaviour the same as before
- it means the API integration page is no longer weirdly showing one-off
  messages