Allow Notify service to send international sms
Right now Notify restricts you to registering with a UK mobile number.
This is because when we built the user registration stuff we couldn’t
send to international mobiles.

However we can send to international mobile numbers, and it’s totally
reasonable to expect employees of the UK government to be working
abroad, and have a foreign mobile phone – we’ve heard from one such
user.

In order for users of Notify to register with an international phone
number, the Notify service needs to have the `international_sms`
permission set. Which this service does, as a data migration.